// generated/src/app/api-models/smparalleljob-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmparalleljobS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  paralleljobIsselfdeleting: number;
  paralleljobLastsavetime: Date;
  paralleljobState: number;
  paralleljobFinishstate: number;
  paralleljobOemjobid: number;
}